Fixed comma errors.

+++ ferdelance/database/tables.py
@@ -46,7 +46,7 @@ class ClientToken(Base):
 
     token_id = Column(Integer, primary_key=True, autoincrement=True, index=True)
     creation_time = Column(DateTime(timezone=True), server_default=now())
-    expiration_time = Column(Integer, nullable=True)
+    expiration_time = Column(Float, nullable=True)
     token = Column(String, nullable=False, index=True, unique=True)
     valid = Column(Boolean, default=True)
